Update API to version 5.1

+++ src/payloads/set_chat_administrator_custom_title.rs
@@ -13,7 +13,7 @@ impl_payload! {
             /// Unique identifier for the target chat or username of the target channel (in the format `@channelusername`)
             pub chat_id: ChatId [into],
             /// Unique identifier of the target user
-            pub user_id: i32,
+            pub user_id: i64,
             /// New custom title for the administrator; 0-16 characters, emoji are not allowed
             pub custom_title: String [into],
         }
